Data-structures reworked, PH-generation recreated
I have decided that policyholders and their statistics-per-period will be lumped into a policy-holder-database. This takes the place of the subgroup class and also takes the functionality of the "stuff" classes that were meant to store PH data during a simulation.

Policy-holder generation now takes place one step at a time, with separate functions to edit a certain field for all policyholders. There currently seems to be a bug with generating claim values.